add 299

+++ leetcode/299_Bulls_and_Cows.hpp
@@ -8,7 +8,7 @@ public:
         int A = 0, B = 0;
         vector<int> ds(10, 0);
         vector<int> gs(10, 0);
-        for (int i = 0; i < secret.length(); ++i) {
+        for (size_t i = 0; i < secret.length(); ++i) {
             if (secret[i] == guess[i]) {
                 ++A;
             } else {
